+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -220,6 +220,11 @@ New checks
   Finds calls to STL library iterator algorithms that could be replaced with
   LLVM range-based algorithms from ``llvm/ADT/STLExtras.h``.
 
+- New :doc:`misc-include-correctness
+  <clang-tidy/checks/misc/include-correctness>` check.
+
+  FIXME: Write a short description.
+
 - New :doc:`misc-override-with-different-visibility
   <clang-tidy/checks/misc/override-with-different-visibility>` check.